+++ admin/loai-hang/list.php
@@ -25,6 +25,7 @@
                     </button>
                 </div>
             </div>
+            <?= $msg ?>
             <div class="activity-categories">
                 <div class="title-categories">
                     <div class="title-pro stt">STT</div>
@@ -43,7 +44,7 @@
                                 <a href="./index.php?btn_edit&ma_loai=<?= $ma_loai ?>">
                                     <button class="btn btn-update updateBtn">Sửa</button>
                                 </a>
-                                <a href="./index.php?btn_delete&ma_loai=<?= $ma_loai ?>">
+                                <a onclick="confirm('Bạn có chắc chắn muốn xóa không')" href="./index.php?btn_delete&ma_loai=<?= $ma_loai ?>">
                                     <button class="btn btn-delete" name="delete">Xóa</button>
                                 </a>
                             </div>
@@ -54,3 +55,9 @@
         </div>
     </div>
 </div>
+<script src="<?= $CONTENT_URL ?>/js/jquery-3.6.0.min.js"></script>
+<script>
+    $(document).ready(function() {
+        $('.alert').delay(3000).slideUp(400, 'linear');
+    });
+</script>
